nizami-backend-master/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
# aspose-words==25.9.0
aspose-words # @ file:///home/anjd/Downloads/aspose_words-25.1.0-py3-none-manylinux1_x86_64.whl
async-timeout==4.0.3
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.18
django-allauth==65.3.1
django-cors-headers==4.6.0
django-environ==0.12.0
django-extensions==3.2.3
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
djangorestframework-datatables==0.7.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.1
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
lxml==5.3.0
marshmallow==3.25.1
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.59.9
orjson==3.10.15
ormsgpack==1.10.0
packaging==24.2
paramiko==3.5.0
pgvector==0.2.5
pillow==11.1.0
propcache==0.2.1
psycopg==3.2.4
psycopg-pool==3.2.4
psycopg2==2.9.10
py==1.4.22
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydevd-pycharm==252.21735.39
PyJWT==2.10.1
PyMuPDF==1.25.2
PyNaCl==1.5.0
pytest==2.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
redis==3.5.3
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uno==0.3.3
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
